+++ src/main/java/com/rolerandomizer/RoleRandomizerConfig.java
@@ -40,17 +40,7 @@ public interface RoleRandomizerConfig extends Config
     {
         return false;
     }
-    
-    @ConfigItem(
-            keyName = "keepPreviousPreferences",
-            name = "Keep role preferences ticked",
-            description = "Keep role preferences ticked after the randomize button has been pressed"
-    )
-    default boolean keepPreferences()
-    {
-        return true;
-    }
-    
+
     @ConfigItem(
             keyName = "addToRando",
             name = "Add to rando",